requirements.txt
asgiref==3.2.10
Django==3.0.8
django-filter==2.3.0
djangorestframework==3.11.1
djangorestframework-jwt==1.11.0
factory-boy==3.0.1
Faker==4.1.1
flake8==3.8.3
freezegun==0.3.15
importlib-metadata==1.7.0
mccabe==0.6.1
psycopg2==2.8.5
pycodestyle==2.6.0
pyflakes==2.2.0
PyJWT==1.7.1
python-dateutil==2.8.1
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
text-unidecode==1.3
zipp==3.1.0
